HELPDESK/requirements.txt
django==4.2
djangorestframework
mssql-django
djangorestframework-simplejwt
django-cors-headers


